need to adjust pic and hedge cases

+++ include/httpRequest.hpp
@@ -6,7 +6,7 @@
 /*   By: jvorstma <jvorstma@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/06/01 17:54:54 by jvorstma      #+#    #+#                 */
-/*   Updated: 2024/08/01 20:25:24 by ibehluli      ########   odam.nl         */
+/*   Updated: 2024/08/02 17:46:09 by ibehluli      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,14 +81,14 @@ class Request {
 		string				getHeaderValue(const string& key) const;
 		// -----------------------
 		void				setUploadeFile(string uploadedFile);
-		void				setBytesCopied(int bytesCopied);
-		void				setMaxLengthUploadContent(int maxLengthUploadContent);
+		void				setBytesCopied(long bytesCopied);
+		void				setMaxLengthUploadContent(long maxLengthUploadContent);
 		void				setBoundary(string const boundary);
 		void				setContentUploadFile(string const contentUploadFile);
 		string 				getBoundary() const;
 		string 				getContentUploadFile() const;
-		int 				getMaxLengthUploadContent();
-		int 				getBytesCopied();
+		long 				getMaxLengthUploadContent();
+		long 				getBytesCopied();
 		string				getUploadedFile() const;
 		// ------------------------
 };
